Move FakeSameThreadExecutorService into core textFixtures
